requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
onnxruntime>=1.15.0 # Or onnxruntime-gpu if GPU is intended and compatible
mss>=9.0.0
pyyaml>=6.0
ultralytics>=8.0.0 # Optional: For easy access to YOLO models/utils if needed later, though we'll use ONNX directly 